Check if user is authorized to update/delete object

+++ favourite/views.py
@@ -22,10 +22,13 @@ class FavouriteView(ViewSet):
     def create(self, request):
         serializer = FavouriteSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
-        serializer.save(user=request.user)
+        serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     def destroy(self, request, pk):
         favourite = get_object_or_404(Favourite, pk=pk)
+        if favourite.user != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         favourite.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
